US004 review changes
